requirements.txt
requests
beautifulsoup4
lxml
google-genai
tkVideoPlayer
moviepy==1.0.3
groq
edge-tts
openai-whisper
yt-dlp
torch
